Changes code licensing from GNU GPL v3 to Apache v2.0
I care about freedom in the broad sense. To me the GNU GPL license
focuses on user freedom a little too narrowly. In doing so it curtails
the freedoms of developers or other non "end users".

At this time I believe the Apache 2.0 license strikes a balance of freedom for
developers and users.