Updates CSP
Content Security Policy appears to be poorly thought through. Basic functions of a modern web such as embedding are fragile as a result of poorly documented and unstable CSP requirements. Resorting to 'unsafe-inline' removes most XSS protections but appears it is required if wishing to avoid a significant amount of work and possible future breakages.